+++ jsonc/testdata/test262/JSON/parse/15.12.1.1-g6-5.js
@@ -5,7 +5,7 @@
 /*---
 es5id: 15.12.1.1-g6-5
 description: >
-    The JSON lexical grammer allows 'n' as a JSONEscapeCharacter after
+    The JSON lexical grammar allows 'n' as a JSONEscapeCharacter after
     '' in a JSONString
 ---*/